+++ CHANGELOG.md
@@ -3,6 +3,30 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com) and this project adheres to [Semantic Versioning](https://semver.org).
 
+## 2.0.1 - 2025-01-22
+
+### Added
+
+- getMainProcess()
+- getSubProcess()
+- getSubProcessEvent()
+
+### Changed
+
+- citronel-error-catalogue
+
+### Deprecated
+
+- Nothing
+
+### Removed
+
+- Nothing
+
+### Fixed
+
+- Nothing
+
 ## 2.0.0 - 2025-01-22
 
 ### Added